+++ utils/signals/interrupts.cpp
@@ -35,14 +35,21 @@ extern "C" {
 #include <unistd.h>
 }
 
+#include <atomic>
+#include <condition_variable>
 #include <cstdlib>
-#include <cstring>
+#include <iostream>
+#include <mutex>
 #include <set>
+#include <thread>
 
+#include "utils/format/macros.hpp"
 #include "utils/logging/macros.hpp"
+#include "utils/noncopyable.hpp"
 #include "utils/process/operations.hpp"
 #include "utils/sanity.hpp"
 #include "utils/signals/exceptions.hpp"
+#include "utils/signals/misc.hpp"
 #include "utils/signals/programmer.hpp"
 
 namespace signals = utils::signals;
@@ -53,200 +60,114 @@ namespace {
 
 
 /// The interrupt signal that fired, or -1 if none.
-static volatile int fired_signal = -1;
+static std::atomic_int fired_signal(-1);
 
+/// Counter for the number of times our signal handler ran.
+static std::atomic_int fired_signal_count(0);
 
-/// Collection of PIDs.
-typedef std::set< pid_t > pids_set;
-
-
-/// List of processes to kill upon reception of a signal.
-static pids_set pids_to_kill;
-
-
-/// Programmer status for the SIGHUP signal.
-static std::auto_ptr< signals::programmer > sighup_handler;
-/// Programmer status for the SIGINT signal.
-static std::auto_ptr< signals::programmer > sigint_handler;
-/// Programmer status for the SIGTERM signal.
-static std::auto_ptr< signals::programmer > sigterm_handler;
-
+/// Cause the signals thread to redeliver the signal to self to terminate.
+///
+/// This is set to true by signals::redeliver_to_exit() before sending the given
+/// signal to self, which the signal thread later uses to forcibly terminate the
+/// program.
+static std::atomic_bool die_now(false);
 
 /// Signal mask to restore after exiting a signal inhibited section.
 static sigset_t global_old_sigmask;
 
+/// Global mutex to protect the rest of the variables below.
+static std::mutex mutex;
 
-/// Whether there is an interrupts_handler object in existence or not.
-static bool interrupts_handler_active = false;
+/// Set to true once the signals thread has finished setting up the handlers.
+static bool started = false;
 
+/// Set to true after the signals thread has finished killing all subprocesses.
+static bool killed = false;
 
-/// Whether there is an interrupts_inhibiter object in existence or not.
-static std::size_t interrupts_inhibiter_active = 0;
+/// Condition variable to wait for started or killed to be set to true.
+static std::condition_variable cv;
+
+/// List of processes to kill upon reception of a signal.
+static std::set< pid_t > pids_to_kill;
 
 
 /// Generic handler to capture interrupt signals.
-///
-/// From this handler, we record that an interrupt has happened so that
-/// check_interrupt() can know whether there execution has to be stopped or not.
-/// We also terminate any of our child processes (started by the
-/// utils::process::children class) so that any ongoing wait(2) system calls
-/// terminate.
 ///
 /// \param signo The signal that caused this handler to be called.
 static void
 signal_handler(const int signo)
 {
-    static const char* message = "[-- Signal caught; please wait for "
-        "cleanup --]\n";
-    if (::write(STDERR_FILENO, message, std::strlen(message)) == -1) {
-        // We are exiting: the message printed here is only for informational
-        // purposes.  If we fail to print it (which probably means something
-        // is really bad), there is not much we can do within the signal
-        // handler, so just ignore this.
-    }
-
     fired_signal = signo;
-
-    for (pids_set::const_iterator iter = pids_to_kill.begin();
-        iter != pids_to_kill.end(); ++iter) {
-        process::terminate_group(*iter);
-    }
+    fired_signal_count += 1;
 }
 
 
-/// Installs signal handlers for potential interrupts.
+/// Unique thread for signal handling.
 ///
-/// \pre Must not have been called before.
-/// \post The various sig*_handler global variables are atomically updated.
+/// This thread must be started with all signals disabled to ensure that it is
+/// the only one reenabling signal handling.
+///
+/// The behavior of this thread is as follows: first, we await for the delivery
+/// of a signal.  Once we get one, we terminate any pending processes which
+/// should cause other threads to get unblocked.  Then we sleep again, waiting
+/// for another signal to terminate ourselves, which can either come from Kyua's
+/// main thread after catching signals::interrupted_error or from a second
+/// delivery of a signal by the user.
 static void
-setup_handlers(void)
+signals_handling_thread(void)
 {
-    PRE(sighup_handler.get() == NULL);
-    PRE(sigint_handler.get() == NULL);
-    PRE(sigterm_handler.get() == NULL);
+    signals::programmer sighup_handler(SIGHUP, signal_handler);
+    signals::programmer sigint_handler(SIGINT, signal_handler);
+    signals::programmer sigterm_handler(SIGTERM, signal_handler);
 
-    // Create the handlers on the stack first so that, if any of them fails, the
-    // stack unwinding cleans things up.
-    std::auto_ptr< signals::programmer > tmp_sighup_handler(
-        new signals::programmer(SIGHUP, signal_handler));
-    std::auto_ptr< signals::programmer > tmp_sigint_handler(
-        new signals::programmer(SIGINT, signal_handler));
-    std::auto_ptr< signals::programmer > tmp_sigterm_handler(
-        new signals::programmer(SIGTERM, signal_handler));
-
-    // Now, update the global pointers, which is an operation that cannot fail.
-    sighup_handler = tmp_sighup_handler;
-    sigint_handler = tmp_sigint_handler;
-    sigterm_handler = tmp_sigterm_handler;
-}
-
-
-/// Uninstalls the signal handlers installed by setup_handlers().
-static void
-cleanup_handlers(void)
-{
-    sighup_handler->unprogram(); sighup_handler.reset(NULL);
-    sigint_handler->unprogram(); sigint_handler.reset(NULL);
-    sigterm_handler->unprogram(); sigterm_handler.reset(NULL);
-}
-
-
-
-/// Masks the signals installed by setup_handlers().
-static void
-mask_signals(sigset_t* old_sigmask)
-{
-    sigset_t mask;
+    ::sigset_t mask, old_mask;
     sigemptyset(&mask);
-    sigaddset(&mask, SIGHUP);
-    sigaddset(&mask, SIGINT);
-    sigaddset(&mask, SIGTERM);
-    const int ret = ::sigprocmask(SIG_BLOCK, &mask, old_sigmask);
+    const int ret = ::pthread_sigmask(SIG_SETMASK, &mask, &old_mask);
     INV(ret != -1);
-}
+#if !defined(NDEBUG)
+    for (int signo = 1; signo < signals::last_signo; ++signo) {
+        if (signo == SIGKILL || signo == SIGSTOP)
+            continue;
+        INV_MSG(sigismember(&old_mask, signo),
+                F("Signal %s not blocked at start of thread") % signo);
+    }
+#endif
 
+    {
+        std::unique_lock< std::mutex > lock(mutex);
+        started = true;
+        cv.notify_all();
+    }
 
-/// Resets the signal masking put in place by mask_signals().
-static void
-unmask_signals(sigset_t* old_sigmask)
-{
-    const int ret = ::sigprocmask(SIG_SETMASK, old_sigmask, NULL);
-    INV(ret != -1);
+    while (fired_signal == -1)
+        (void)::sigsuspend(&mask);
+    std::cerr << "[-- Signal caught; please wait for cleanup --]\n";
+
+    {
+        std::unique_lock< std::mutex > lock(mutex);
+        for (pid_t pid : pids_to_kill) {
+            process::terminate_group(pid);
+        }
+        killed = true;
+        cv.notify_all();
+    }
+
+    while (!die_now && fired_signal_count == 1)
+        (void)::sigsuspend(&mask);
+    if (!die_now) {
+        // Only print the message if the second signal is not because of
+        // signals::redeliver_on_exit(), which would indicate a controlled exit
+        // from the main thread.
+        std::cerr << "[-- Double signal caught; terminating --]\n";
+    }
+    sigterm_handler.unprogram();
+    sigint_handler.unprogram();
+    sighup_handler.unprogram();
+    ::kill(::getpid(), fired_signal);
 }
 
 
 }  // anonymous namespace
-
-
-/// Constructor that sets up the signal handlers.
-signals::interrupts_handler::interrupts_handler(void) :
-    _programmed(false)
-{
-    PRE(!interrupts_handler_active);
-    setup_handlers();
-    _programmed = true;
-    interrupts_handler_active = true;
-}
-
-
-/// Destructor that removes the signal handlers.
-///
-/// Given that this is a destructor and it can't report errors back to the
-/// caller, the caller must attempt to call unprogram() on its own.
-signals::interrupts_handler::~interrupts_handler(void)
-{
-    if (_programmed) {
-        LW("Destroying still-programmed signals::interrupts_handler object");
-        try {
-            unprogram();
-        } catch (const error& e) {
-            UNREACHABLE;
-        }
-    }
-}
-
-
-/// Unprograms all signals captured by the interrupts handler.
-///
-/// \throw system_error If the unprogramming of any signal fails.
-void
-signals::interrupts_handler::unprogram(void)
-{
-    PRE(_programmed);
-
-    // Modify the control variables first before unprogramming the handlers.  If
-    // we fail to do the latter, we do not want to try again because we will not
-    // succeed (and we'll cause a crash due to failed preconditions).
-    _programmed = false;
-    interrupts_handler_active = false;
-
-    cleanup_handlers();
-    fired_signal = -1;
-}
-
-
-/// Constructor that sets up signal masking.
-signals::interrupts_inhibiter::interrupts_inhibiter(void)
-{
-    sigset_t old_sigmask;
-    mask_signals(&old_sigmask);
-    if (interrupts_inhibiter_active == 0) {
-        global_old_sigmask = old_sigmask;
-    }
-    ++interrupts_inhibiter_active;
-}
-
-
-/// Destructor that removes signal masking.
-signals::interrupts_inhibiter::~interrupts_inhibiter(void)
-{
-    if (interrupts_inhibiter_active > 1) {
-        --interrupts_inhibiter_active;
-    } else {
-        interrupts_inhibiter_active = false;
-        unmask_signals(&global_old_sigmask);
-    }
-}
 
 
 /// Checks if an interrupt has fired.
@@ -262,8 +183,10 @@ signals::interrupts_inhibiter::~interrupts_inhibiter(void)
 void
 signals::check_interrupt(void)
 {
-    if (fired_signal != -1) {
-        const int original_fired_signal = fired_signal;
+    const int original_fired_signal = fired_signal;
+    if (original_fired_signal != -1) {
+        std::unique_lock< std::mutex > lock(mutex);
+        cv.wait(lock, []{return killed;});
         fired_signal = -1;
         throw interrupted_error(original_fired_signal);
     }
@@ -272,14 +195,14 @@ signals::check_interrupt(void)
 
 /// Registers a child process to be killed upon reception of an interrupt.
 ///
-/// \pre Must be called with interrupts being inhibited.  The caller must ensure
-/// that the call call to fork() and the addition of the PID happen atomically.
+/// \pre The caller must ensure that the call to fork() and the addition of the
+/// PID happen without interrupts checking in between.
 ///
-/// \param pid The PID of the child process.  Must not have been yet regsitered.
+/// \param pid The PID of the child process.  Must not have been yet registered.
 void
 signals::add_pid_to_kill(const pid_t pid)
 {
-    PRE(interrupts_inhibiter_active);
+    std::lock_guard< std::mutex > guard(mutex);
     PRE(pids_to_kill.find(pid) == pids_to_kill.end());
     pids_to_kill.insert(pid);
 }
@@ -287,16 +210,67 @@ signals::add_pid_to_kill(const pid_t pid)
 
 /// Unregisters a child process previously registered via add_pid_to_kill().
 ///
-/// \pre Must be called with interrupts being inhibited.  This is not necessary,
-/// but pushing this to the caller simplifies our logic and provides consistency
-/// with the add_pid_to_kill() call.
-///
 /// \param pid The PID of the child process.  Must have been registered
 ///     previously, and the process must have already been awaited for.
 void
 signals::remove_pid_to_kill(const pid_t pid)
 {
-    PRE(interrupts_inhibiter_active);
+    std::lock_guard< std::mutex > guard(mutex);
     PRE(pids_to_kill.find(pid) != pids_to_kill.end());
     pids_to_kill.erase(pid);
+}
+
+
+/// Starts the signals handling thread to handle interrupts asynchronously.
+///
+/// This configures the program to funnel all signal handling through a single
+/// thread, started here.  All other threads must then check for interrupts in
+/// strategic places by invoking check_interrupt().
+///
+/// This should be called early in the main thread, before any other threads
+/// have been started, to ensure the right default signal mask is set for them.
+void
+signals::setup_interrupts(void)
+{
+    PRE(!started);
+
+    ::sigset_t mask;
+    sigfillset(&mask);
+    const int ret = ::sigprocmask(SIG_BLOCK, &mask, &global_old_sigmask);
+    INV(ret != -1);
+
+    std::thread thread(signals_handling_thread);
+    thread.detach();
+
+    // Wait until the thread finishes starting up and configuring signal
+    // handling.  This is necessary to avoid losing interrupts if
+    // check_interrupt() is called too soon after the thread starts.
+    std::unique_lock< std::mutex > lock(mutex);
+    cv.wait(lock, []{return started;});
+}
+
+
+/// Clears interrupts handling in a new child.
+///
+/// This must be invoked right after fork() to ensure the child process can
+/// receive signals.
+void
+signals::reset_interrupts_in_new_child(void)
+{
+    const int ret = ::sigprocmask(SIG_SETMASK, &global_old_sigmask, NULL);
+    INV(ret != -1);
+}
+
+
+/// Redeliver a caught signal to cause the program to terminate.
+///
+/// This has to be invoked from the main thread once we have caught
+/// interrupted_error to cause the program to terminate with the right exit
+/// status.
+void
+signals::redeliver_to_exit(const int signo)
+{
+    die_now = true;
+    ::kill(::getpid(), signo);
+    LD("Interrupt signal re-delivery did not terminate program");
 }